# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: TechCV Profile API
  version: 0.1.0
  description: Profile endpoints extracted from techcv-design/openapi.yaml
servers:
  - url: /api
    description: Relative API base path
security:
  - SessionAuth: []
  - BearerAuth: []
tags:
  - name: Profile
paths:
  /profile:
    get:
      tags:
        - Profile
      summary: Get profile
      operationId: getProfile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags:
        - Profile
      summary: Update profile
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateProfileResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    VisibilitySettings:
      type: object
      additionalProperties:
        type: boolean
    Profile:
      type: object
      required:
        - id
        - userId
        - visibilitySettings
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        userId:
          type: string
        fullName:
          type: string
        nickname:
          type: string
        location:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        summary:
          type: string
        githubUrl:
          type: string
          format: uri
        zennUrl:
          type: string
          format: uri
        qiitaUrl:
          type: string
          format: uri
        websiteUrl:
          type: string
          format: uri
        preferredWorkStyle:
          type: string
        visibilitySettings:
          $ref: "#/components/schemas/VisibilitySettings"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GetProfileResponse:
      type: object
      required:
        - profile
      properties:
        profile:
          $ref: "#/components/schemas/Profile"
    UpdateProfileResponse:
      type: object
      required:
        - profile
      properties:
        profile:
          $ref: "#/components/schemas/Profile"
    ProfileUpdateRequest:
      type: object
      properties:
        fullName:
          type: string
        nickname:
          type: string
        location:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        summary:
          type: string
        githubUrl:
          type: string
          format: uri
        zennUrl:
          type: string
          format: uri
        qiitaUrl:
          type: string
          format: uri
        websiteUrl:
          type: string
          format: uri
        preferredWorkStyle:
          type: string
        visibilitySettings:
          $ref: "#/components/schemas/VisibilitySettings"